Make game functions work
Game now works. Changed the even listners for the game buttons to be delegated from the parent instead of running a forEach on the buttons. Not entirely sure why the forEach wasn't working, but the delegation fixed the issue. Started work on end game. Code is working so far, just need to create an end screen in the UI.